Use the ParseFromBytes trait to parse messages
In order to represent the different kinds of spectrum analyzer sweeps
with a single Sweep type, we need to distinguish between types that can
be parsed from bytes and types that are distinct RF Explorer messages
with their own message prefix. This commit allows us to use the same
code to parse types like Sweep and true RF Explorer messages like
Config.